requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
pymongo==3.9.0
PyYAML==5.1.2
requests==2.22.0
selenium==3.141.0
tqdm==4.36.0
urllib3==1.25.3
